Remove debugging code.
